Balancing Values
easy levels base score x 1.5 so its easier compared to x 2

Change scores to float value

+++ Monster/Assets/Scripts/GameManagerScript/ManagerScript/LevelManagerScriptableObject.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 public class LevelManagerScriptableObject : ScriptableObject
 {
     public int buildingsInScene;
-    public int baseScore;
+    public float baseScore;
     public float destructionLevel;
     public int cityLevel;
     public float currentDestruction;
